Split detectors into four modules so the change is reviewable
Software Factory kept reporting that functions in this PR were not
implemented, citing line numbers that resolve to different code on main. My
diagnosis was right about the mechanism and wrong about what to do with it.
Its window into a changed file is positional, and I treated that as its
problem to fix. It is also mine: a 1,948-line module whose entire new surface
sits in the back half is hard to review, by a person or by a tool, and
"technically it is all there" is not a standard worth defending.

detectors.py 1,948 -> 533 lines, four new modules, none of them a dumping
ground:

  detector_calibration.py   where a threshold comes from, in four layers
  detector_core.py          parse, normalize, redact: the shared floor
  detector_behaviour.py     the four behavioural detectors and their patterns
  detector_money.py         what a finding costs, and what may escalate it

detectors.py keeps the four trajectory detectors, the registry and run_all,
and re-exports the rest, so clawmetry.detectors remains the single import for
every caller. The seams are the ones that were already in the file as section
comments; nothing was rewritten in the move.

No test changed. All 110 pass, which is the point: a behaviour-preserving
move should not need the tests adjusted to agree with it.
