+++ CLAUDE.md
@@ -53,7 +53,11 @@ All HTTP endpoints live here, organised by feature. Each module owns one or more
 | `clawmetry/local_store.py` | ~12,000 | **DuckDB store** — the single data layer features read/write (daemon holds the writer lock) |
 | `clawmetry/local_server.py` | ~200 | Daemon-hosted localhost query server (`/__local_query__/<method>`) so the dashboard/sync read DuckDB without grabbing the writer lock |
 | `clawmetry/proxy.py` | ~2,700 | Enforcement proxy — budget limits, loop detection, model routing (port 4100) |
-| `clawmetry/detectors.py` | ~1,790 | Judge-free detectors over a session's tool stream, run by the daemon each tick. **Trajectory** (is it stuck?): `stuck_loop`, `no_progress`, `repeated_tool_failure`, `action_discrepancy`. **Behavioural** (is it doing something it does not normally do?): `file_blast_radius`, `credential_access`, `network_egress`, `privilege_change`. Also owns per-runtime threshold calibration (`resolve_thresholds`, `RUNTIME_PROFILES`) and the spend-at-risk model (`annotate_spend`) |
+| `clawmetry/detectors.py` | ~530 | The eight detectors' registry and orchestration. `run_all` normalizes once and shares the parse with every detector, then returns incidents ranked by what ignoring them costs. Holds the four **trajectory** detectors (is it stuck?): `stuck_loop`, `no_progress`, `repeated_tool_failure`, `action_discrepancy`. Re-exports the four modules below, so `clawmetry.detectors` stays the single import |
+| `clawmetry/detector_core.py` | ~670 | Parse, normalize, redact. The shared floor: `normalize_events`, the action surface (paths / command / hosts, heredoc bodies stripped), and the redaction rules that decide what an incident may publish |
+| `clawmetry/detector_behaviour.py` | ~470 | The four **behavioural** detectors (is it doing something it does not normally do?): `file_blast_radius`, `credential_access`, `network_egress`, `privilege_change`, with their pattern tables |
+| `clawmetry/detector_calibration.py` | ~280 | Where a threshold comes from: module defaults → `RUNTIME_PROFILES` → the cohort's learned baseline → a per-runtime env override. `resolve_thresholds` reports which layer set each value |
+| `clawmetry/detector_money.py` | ~130 | `spend_at_risk_usd` and the ranking. Only a measured basis may promote a warning to `critical` |
 | `clawmetry/interceptor.py` | ~630 | Zero-config HTTP monkey-patching for LLM cost tracking (patches httpx/requests) |
 | `clawmetry/providers_pricing.py` | ~430 | Multi-provider pricing table (Anthropic, OpenAI, Google, OpenRouter, etc.) |
 | `clawmetry/config.py` | ~200 | Configuration dataclass |
